Unequip and Equip get called multiple times during startup
In the start scene, equip and unequip methods are called multiple times in order to show the model in the character select screen.

I don't want my backpack operations running in the start scene since a large portion of the functionality isn't initialized yet.

This is what was causing the error.  This now detects for start scene and ignores backpack functions.